# server/docs/swagger.yaml
openapi: 3.0.0
info:
  title: TradeVault API
  description: |
    A comprehensive trading platform API that provides:

    ## Features
    - **User Authentication**: Secure login/register with JWT tokens
    - **OTP Verification**: Email-based OTP system for account security
    - **Real-time Trading**: WebSocket-based live stock updates
    - **Stock Management**: Buy/sell stocks with portfolio tracking
    - **Biometric Authentication**: Advanced security with biometric keys
    - **OAuth Integration**: Google and Apple sign-in support

    ## Authentication
    - **App Token**: Used for main application endpoints (JWT_SECRET)
    - **Socket Token**: Used for WebSocket connections (SOCKET_TOKEN_SECRET)

    ## Rate Limiting
    - OTP requests: 1 per minute per email
    - Login attempts: 3 attempts before 30-minute lockout
    - PIN attempts: 3 attempts before 30-minute lockout

  version: "1.0.0"
  contact:
    name: TradeVault Support
    email: support@tradevault.com
    url: https://tradevault.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.tradevault.com
    description: Production server
security:
  - BearerAuth: []
  - SocketAuth: []
paths:
...
  /auth/refresh-token:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Refreshes the access token using a valid refresh token.
        Supports both app and socket token refresh.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
            example:
              type: "app"
              refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        "400":
          description: Bad request - invalid refresh type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ========================================
  # USER MANAGEMENT ENDPOINTS (Protected)
  # ========================================
  /auth/profile:
    get:
      tags:
        - User Management
      summary: Get user profile
      description: Retrieves the authenticated user's profile information
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - User Management
      summary: Update user profile
      description: Updates the authenticated user's profile information
      operationId: updateProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
            example:
              name: "John Doe"
              gender: "male"
              date_of_birth: "1990-01-01"
              phone_number: "1234567890"
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateProfileResponse"
        "400":
          description: Bad request - invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/verify-pin:
    post:
      tags:
        - User Management
      summary: Verify login PIN
      description: Verifies the user's login PIN and returns socket tokens
      operationId: verifyPin
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPinRequest"
            example:
              login_pin: "1234"
      responses:
        "200":
          description: PIN verified successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyPinResponse"
        "400":
          description: Bad request - PIN not set or invalid format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - invalid PIN or account locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /stocks/stock:
    get:
      tags:
        - Stock Management
      summary: Get stock by symbol
      description: Retrieves detailed information about a specific stock
      operationId: getStockBySymbol
      security:
        - BearerAuth: []
      parameters:
        - name: stock
          in: query
          description: Stock symbol (e.g., AAPL, GOOGL)
          required: true
          schema:
            type: string
          example: "AAPL"
      responses:
        "200":
          description: Stock retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetStockResponse"
        "400":
          description: Bad request - missing stock symbol
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Stock not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /stocks/buy:
    post:
      tags:
        - Trading
      summary: Buy stock
      description: Purchases a specified quantity of a stock
      operationId: buyStock
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyStockRequest"
            example:
              stock_id: "507f1f77bcf86cd799439011"
              quantity: 10
      responses:
        "201":
          description: Stock purchased successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuyStockResponse"
        "400":
          description: Bad request - insufficient balance or invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stocks/sell:
    post:
      tags:
        - Trading
      summary: Sell stock
      description: Sells a specified quantity of a stock from holdings
      operationId: sellStock
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellStockRequest"
            example:
              holdingId: "507f1f77bcf86cd799439012"
              quantity: 5
      responses:
        "200":
          description: Stock sold successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellStockResponse"
        "400":
          description: Bad request - insufficient quantity or invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stocks/holding:
    get:
      tags:
        - Trading
      summary: Get user holdings
      description: Retrieves all stock holdings for the authenticated user
      operationId: getAllHoldings
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Holdings retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetHoldingsResponse"
        "401":
          description: Unauthorized - invalid socket token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for application authentication.
        Include in Authorization header: `Bearer <token>`
    SocketAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for WebSocket and trading operations.
        Include in Authorization header: `Bearer <token>`
  schemas:
...
    RefreshTokenRequest:
      type: object
      required:
        - type
        - refresh_token
      properties:
        type:
          type: string
          enum: [socket, app]
          description: Type of token to refresh
          example: "app"
        refresh_token:
          type: string
          description: Valid refresh token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    RefreshTokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: New JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          description: New JWT refresh token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    # ========================================
    # USER MANAGEMENT SCHEMAS
    # ========================================
    ProfileResponse:
      type: object
      properties:
        userId:
          type: string
          description: Unique user identifier
          example: "507f1f77bcf86cd799439011"
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        phone_exist:
          type: boolean
          description: Whether user has set a phone number
          example: false
        name:
          type: string
          description: User's display name
          example: "John Doe"
        login_pin_exist:
          type: boolean
          description: Whether user has set a login PIN
          example: false
        balance:
          type: string
          description: User's account balance (formatted to 2 decimal places)
          example: "50000.00"
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 50
          description: User's display name
          example: "John Doe"
        gender:
          type: string
          enum: [male, female, other]
          description: User's gender
          example: "male"
        date_of_birth:
          type: string
          format: date
          description: User's date of birth
          example: "1990-01-01"
    UpdateProfileResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Operation success status
          example: true
        data:
          $ref: "#/components/schemas/User"
...
    VerifyPinRequest:
      type: object
      required:
        - login_pin
      properties:
        login_pin:
          type: string
          pattern: "^[0-9]{4}$"
          description: 4-digit login PIN
          example: "1234"
    VerifyPinResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Operation success status
          example: true
        socket_tokens:
          type: object
          properties:
            socket_access_token:
              type: string
              description: JWT token for WebSocket authentication
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            socket_refresh_token:
              type: string
              description: JWT refresh token for WebSocket
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...
    GetStockResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "Stock retrieved successfully!"
        data:
          $ref: "#/components/schemas/Stock"

    # ========================================
    # TRADING SCHEMAS
    # ========================================
    BuyStockRequest:
      type: object
      required:
        - stock_id
        - quantity
      properties:
        stock_id:
          type: string
          description: Stock identifier
          example: "507f1f77bcf86cd799439011"
        quantity:
          type: integer
          minimum: 1
          description: Number of shares to buy
          example: 10
    BuyStockResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "Stock bought successfully!"
        data:
          $ref: "#/components/schemas/Holding"
    SellStockRequest:
      type: object
      required:
        - holdingId
        - quantity
      properties:
        holdingId:
          type: string
          description: Holding identifier
          example: "507f1f77bcf86cd799439012"
        quantity:
          type: integer
          minimum: 1
          description: Number of shares to sell
          example: 5
    SellStockResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "Stock sold successfully!"
        data:
          type: object
          properties:
            orderId:
              type: string
              description: Order identifier
              example: "507f1f77bcf86cd799439013"
            sellPrice:
              type: number
              description: Total sale amount
              example: 1277.50
    GetHoldingsResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "Holdings retrieved successfully!"
        data:
          type: array
          items:
            $ref: "#/components/schemas/Holding"
...
    User:
      type: object
      properties:
        _id:
          type: string
          description: Unique user identifier
          example: "507f1f77bcf86cd799439011"
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        name:
          type: string
          description: User's display name
          example: "John Doe"
        phone_number:
          type: string
          nullable: true
          description: User's phone number
          example: "+1234567890"
        date_of_birth:
          type: string
          format: date
          nullable: true
          description: User's date of birth
          example: "1990-01-01"
        gender:
          type: string
          enum: [male, female, other]
          nullable: true
          description: User's gender
          example: "male"
        balance:
          type: number
          description: User's account balance
          example: 50000.00
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2024-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-01T00:00:00.000Z"
    Stock:
      type: object
      properties:
        _id:
          type: string
          description: Unique stock identifier
          example: "507f1f77bcf86cd799439011"
        symbol:
          type: string
          description: Stock symbol/ticker
          example: "TSLA"
        companyName:
          type: string
          description: Company name
          example: "Tesla, Inc."
        iconUrl:
          type: string
          format: uri
          description: URL to company icon
          example: "https://example.com/tesla-icon.png"
        lastDayTradedPrice:
          type: string
          description: Previous day's closing price
          example: "250.00"
        currentPrice:
          type: number
          description: Current stock price
          example: 255.50
        dayTimeSeries:
          type: array
          description: Intraday price data
          items:
            type: object
        tenMinTimeSeries:
          type: array
          description: 10-minute interval price data
          items:
            type: object
    Holding:
      type: object
      properties:
        _id:
          type: string
          description: Unique holding identifier
          example: "507f1f77bcf86cd799439011"
        user:
          type: string
          description: User identifier
          example: "507f1f77bcf86cd799439012"
        stock:
          type: string
          description: Stock identifier
          example: "507f1f77bcf86cd799439013"
        quantity:
          type: integer
          description: Number of shares held
          example: 10
        buyPrice:
          type: number
          description: Purchase price per share
          example: 250.00
        createdAt:
          type: string
          format: date-time
          description: Holding creation timestamp
          example: "2024-01-01T00:00:00.000Z"
...
    ErrorResponse:
      type: object
      properties:
        msg:
          type: string
          description: Error message
          example: "Invalid request data"
...